requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-dotenv